+++ projectroles/remote_projects.py
@@ -29,7 +29,6 @@ from projectroles.models import (
     AppSetting,
 )
 from projectroles.plugins import get_backend_api
-from projectroles.utils import build_secret
 
 
 app_settings = AppSettingAPI()
@@ -539,9 +538,8 @@ class RemoteProjectAPI:
                 user=user, email=e
             ).exists():
                 continue
-            # TODO: Remove redundant secret once #1477 is implemented
             email_obj = SODARUserAdditionalEmail.objects.create(
-                user=user, email=e, verified=True, secret=build_secret(16)
+                user=user, email=e, verified=True
             )
             logger.info(
                 'Created user {} additional email "{}"'.format(